username and password

+++ server.py
@@ -14,48 +14,45 @@ class ClientThread(Thread):
         print("[+] New server socket thread started for " + ip + ":" + str(port))
 
     def run(self):
+        data = conn.recv(2048).decode()
+        username = data.split(',')[0]
+        password = data.split(',')[1]
+        if username == "root" and password == "root":
+            conn.send("True".encode())
+            while True:
+                data = conn.recv(2048).decode()
 
-        while True:
-            data = conn.recv(2048).decode()
-            print("Server received data:", data)
-            MessageList = data.split()
-            Order = MessageList[0]
-            if Order == 'Quit':
-                break
-            if Order == 'LIST':
-                ClientThread.list_server_files(self)
-                print("List")
-            if Order == 'RETR':
-                ClientThread.retrive_file(self, MessageList[1])
-                print("retrive")
-            if Order == 'DELE':
-                ClientThread.delete_file(self, MessageList[1])
-                print("delete")
-            if Order == 'RMD':
-                ClientThread.delete_all_cached_files(self)
-                print("RMD")
+                print("Server received data:", data)
+                MessageList = data.split()
+                Order = MessageList[0]
+                if Order == 'Quit':
+                    break
+                if Order == 'LIST':
+                    ClientThread.list_server_files(self)
+                    print("List")
+                if Order == 'RETR':
+                    ClientThread.retrive_file(self, MessageList[1])
+                    print("retrive")
+                if Order == 'DELE':
+                    ClientThread.delete_file(self, MessageList[1])
+                    print("delete")
+                if Order == 'RMD':
+                    ClientThread.delete_all_cached_files(self)
+                    print("RMD")
+        else:
+            conn.send("False".encode())
+
                 # MESSAGE = input("Server: Enter Response from Server/Enter Quit:")
                 # if MESSAGE == 'Quit':
                 #     break
 
                 # conn.send(MESSAGE.encode())  # echo
 
-    def delete_file(self, file_name):
-        local_files_list = self.local_files()
-        if (file_name in local_files_list):
-            os.remove("File/" + file_name)
-        else:
-            print("no such file in cache proxy!")
-
-    def delete_all_cached_files(self):
-        local_files_list = self.local_files()
-        for file in local_files_list:
-            self.delete_file(file)
-
     def retrive_file(self, file_name):
         host = socket.gethostbyname('ceit.aut.ac.ir')
         request = "GET /~94131090/CN1_Project_Files/" + str(file_name) + " HTTP/1.1\r\nHost: " + host + "\r\n\r\n"
         proxy_files = self.local_files()
+        print("going to check")
         for file in proxy_files:
             if file == file_name:
                 print("in local")
@@ -72,7 +69,7 @@ class ClientThread(Thread):
             while True:
                 print('receiving data...')
                 data = ServerSocket.recv(1024)
-                re.sub(b'.*\r\n\r\n', '', data)
+                re.sub(r'.*\n\n', '', data.decode())
                 print('data=%s', (data))
                 if not data:
                     break
@@ -98,7 +95,6 @@ class ClientThread(Thread):
         while len(response) > 0:
             response = ServerSocket.recv(1000).decode()
             result += response
-        print(result)
         urls = re.findall(r'href=[\'"]?([^\'" >]+)', result)
         urls.pop(0)
         urls.pop(0)
@@ -106,23 +102,57 @@ class ClientThread(Thread):
         urls.pop(0)
         urls.pop(0)
         print(urls)
+        ehsan = ','.join(urls)
+        dataConn.send(ehsan.encode())
         ServerSocket.close()
+
+
+    # def return_server_files(self, file_list):
+    #     TCP_IP = '127.0.0.1'
+    #     TCP_PORT = 6021
+    #     BUFFER_SIZE = 20  # Usually 1024, but we need quick response
+    #
+    #     tcpServer = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    #     tcpServer.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+    #     tcpServer.bind((TCP_IP, TCP_PORT))
+
+    def delete_file(self, file_name):
+        local_files_list = self.local_files()
+        if (file_name in local_files_list):
+            os.remove("File/" + file_name)
+        else:
+            print("no such file in cache proxy!")
+
+    def delete_all_cached_files(self):
+        local_files_list = self.local_files()
+        for file in local_files_list:
+            self.delete_file(file)
 
 
 # Multithreaded Python server : TCP Server Socket Program Stub
 TCP_IP = '127.0.0.1'
-TCP_PORT = 6020
+data_port = 3020
+control_port = 3021
+
 BUFFER_SIZE = 20  # Usually 1024, but we need quick response 
 
-tcpServer = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-tcpServer.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-tcpServer.bind((TCP_IP, TCP_PORT))
+controlSocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+controlSocket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+controlSocket.bind((TCP_IP, control_port))
 threads = []
 
+dataSocket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+dataSocket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+dataSocket.bind((TCP_IP, data_port))
+
+
+
 while True:
-    tcpServer.listen(4)
+    controlSocket.listen(4)
+    dataSocket.listen(4)
     print("Waiting for connections from clients...")
-    (conn, (ip, port)) = tcpServer.accept()
+    (conn, (ip, port)) = controlSocket.accept()
+    (dataConn, (dip, Dport)) = dataSocket.accept()
     newthread = ClientThread(ip, port)
     newthread.start()
     threads.append(newthread)
